requirements.txt
azure-core
azure-ai-formrecognizer
azure-storage-blob
streamlit
python-dotenv